Update for newer versions of GDK.
In particular, the icon window was not receiving enter and leave events from
the pointer.  To fix this, we get rid of the second GdkWindow iconwin entirely
and set win to be its own icon.

This also removes the need for the "broken window manager" fix, so we remove it
and all references to it.