Rewrite to use GDK3/Cairo
I migrated this appllication to GTK3 from the old GTK2 routines so that
it would be slightly more modernized. The original code is almost 20
years old, so was very depcreated, and this should future proof it a
little more.

The biggest change necessary was converting the bit-order of the
colormaps that are used to draw the flame effect and the draw_fire()
function for Cairo. Cairo's RGB-24 format needs the most-significant 8
bits of the integer to be unused, with the second-most significant byte
to be the R value, then G and B.